Added comilation support for Google Mock
