+++ src/Igorek.cpp
@@ -1,0 +1,9 @@
+#include "Igorek.h"
+
+int Igorek::mul(int a, int b) {
+    return a * b;
+}
+
+int sum (int num1, int num2) {
+    return num1 + num2;
+}
