+++ src/Igorek.h
@@ -1,6 +1,9 @@
 #ifndef TEST_H
 #define TEST_H
 
+class Igorek {
+    virtual int mul(int a, int b) = 0;
+};
 
 int sum (int num1, int num2);
