--- src/test.cpp
@@ -1,6 +1,0 @@
-#include "test.h"
-
-
-int sum (int num1, int num2) {
-    return num1 + num2;
-}
